refactor: 정적 자원 처리 핸들러 패키지 구조 변경
