+++ tomcat/src/main/java/org/apache/coyote/http11/handle/HttpHandlerMapper.java
@@ -1,11 +1,11 @@
 package org.apache.coyote.http11.handle;
 
 import java.util.Set;
-import org.apache.coyote.http11.handle.handler.CssHttpHandler;
+import org.apache.coyote.http11.handle.handler.resource.CssHttpHandler;
 import org.apache.coyote.http11.handle.handler.DefaultHttpHandler;
-import org.apache.coyote.http11.handle.handler.HtmlHttpHandler;
+import org.apache.coyote.http11.handle.handler.resource.HtmlHttpHandler;
 import org.apache.coyote.http11.handle.handler.HttpHandler;
-import org.apache.coyote.http11.handle.handler.JsHttpHandler;
+import org.apache.coyote.http11.handle.handler.resource.JsHttpHandler;
 import org.apache.coyote.http11.reqeust.HttpRequest;
 
 public class HttpHandlerMapper {
