+++ tomcat/src/main/java/org/apache/coyote/http11/handle/handler/resource/StaticResourceHandler.java
@@ -1,10 +1,11 @@
-package org.apache.coyote.http11.handle.handler;
+package org.apache.coyote.http11.handle.handler.resource;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import org.apache.coyote.http11.handle.handler.HttpHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
